--- core-primitives/types/Cargo.toml
@@ -1,46 +1,0 @@
-[package]
-name = "itp-types"
-version = "0.8.0"
-authors = ["Integritee AG <hello@integritee.network>"]
-edition = "2018"
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
-primitive-types = { version = "0.11.1", default-features = false, features = ["codec"] }
-chrono          = { version = "0.4.19", default-features = false, features = ["alloc"]}
-serde           = { version = "1.0", default-features = false, features = ["alloc", "derive"] }
-serde_json      = { version = "1.0", default-features = false, features = ["alloc"] }
-substrate-api-client = { git = "https://github.com/scs/substrate-api-client", branch = "polkadot-v0.9.20-asset", default-features = false }
-
-sgx_tstd = { branch = "master", features = ["untrusted_fs","net","backtrace"], git = "https://github.com/apache/teaclave-sgx-sdk.git", optional = true}
-
-# local deps
-itp-storage = { path = "../storage", default-features = false }
-
-# substrate-deps
-sp-std = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.20"}
-sp-core = { version = "6.0.0", default-features = false, features = ["full_crypto"], git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.20"}
-sp-finality-grandpa = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.20" }
-sp-runtime = { version = "6.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.20" }
-
-# node deps
-pallet-ajuna-gameregistry  = { default-features = false, git = "https://github.com/ajuna-network/Ajuna", branch = "release/bengal" }
-
-[features]
-default = ['std']
-std = [ 'codec/std',
-        'sp-core/std',
-        'sp-finality-grandpa/std',
-        'chrono/std',
-        'serde/std',
-        'serde_json/std',
-        'pallet-ajuna-gameregistry/std',
-        'primitive-types/std',
-        'itp-storage/std',
-        'sp-runtime/std',
-        'substrate-api-client/std',
-]
-sgx = [ 'sgx_tstd' ]
-
-[dev-dependencies]
-sp-keyring = { version = "6.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.20"}
